complete dependency checks
