--- local_libs/lib-dex-commons/src/main/kotlin/gov/cdc/dex/azure/CosmosDBProxySimple.kt
@@ -1,14 +1,0 @@
-package gov.cdc.dex.azure
-
-import com.azure.cosmos.CosmosClientBuilder
-import com.azure.cosmos.CosmosClient
-
-class CosmosDBProxySimple (serviceEndpoint: String, key: String) {
-    private val cosmosClient: CosmosClient = CosmosClientBuilder()
-        .endpoint(serviceEndpoint)
-        .key(key)
-        .buildClient()
-    fun disconnect() {
-        cosmosClient.close()
-    }
-}
